+++ CHANGELOG.md
@@ -5,6 +5,11 @@
 * [CHANGED] [5495](https://github.com/stripe/stripe-android/pull/5495) Upgrade `compileSdkVersion`
   to 33.
 
+### PaymentSheet
+
+* [ADDED][5502](https://github.com/stripe/stripe-android/pull/5502) Added phone number minimum length
+  validation
+
 ## 20.11.0 - 2022-08-29
 This release adds postal code validation for PaymentSheet and fixed a fileprovider naming bug for Identity.